Update in journal controller file

+++ src/main/java/com/springBoot/journalApp/services/JournalEntryService.java
@@ -13,6 +13,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalDateTime;
 import java.util.*;
+
 @Component
 @Slf4j
 public class JournalEntryService {
@@ -42,7 +43,7 @@ public class JournalEntryService {
         try{
             journalEntryRepo.save(journalEntry);
         } catch(Exception ex){
-            log.error(String.valueOf(ex));
+            log.error("Error eccured of {} ",String.valueOf(ex));
         }
 
     }
@@ -55,11 +56,19 @@ public class JournalEntryService {
         return journalEntryRepo.findById(id);
     }
 
-    public void deleteById(ObjectId id, String userName){
-        User user = userService.findByUserName(userName);
-        user.getJournalEntities().removeIf(x -> x.getId().equals(id));
-        userService.saveEntry(user);
-        journalEntryRepo.deleteById(id);
+    public boolean deleteById(ObjectId id, String userName){
+        boolean removeJournal = false;
+        try {
+            User user = userService.findByUserName(userName);
+            removeJournal = user.getJournalEntities().removeIf(x -> x.getId().equals(id));
+            if(removeJournal) {
+                userService.saveEntry(user);
+                journalEntryRepo.deleteById(id);
+            }
+        }
+        catch (Exception ex){
+            throw new RuntimeException("An Error occured in deleting the entry");
+        }
+        return removeJournal;
     }
-
 }
